feat: add typecsript to moviecard
